refactor(PR-611): Move VA toggle to voice-assistant-chat
