+++ src/app/ui-components/sidebar-right/sidebar-right.component.html
@@ -3,41 +3,6 @@
         <div class="mt-4">
             <div
                 class="row"
-                *ngIf="lStorage === 'chat'">
-                <label class="text-center">Voice interaction</label>
-                <div class="col">
-                    <div class="row mt-1 mb-3">
-                        <div
-                            class="col-11 ml-3 on-off"
-                            [ngClass]="vaState ? 'on' : 'off'">
-                            <div class="row">
-                                <button
-                                    type="button"
-                                    class="col btn text-truncate element text-center p-3"
-                                    id="sidebar-right-toggle-voice-assistant"
-                                    (click)="toggleVoiceAssistant()"
-                                    title="Toggle for turning Voice Assistant on and off"
-                                    value="{{vaState}}">
-                                    <div
-                                        *ngIf="vaState; then thenBlock else elseBlock"></div>
-                                    <ng-template #thenBlock
-                                        ><label class="mb-0"
-                                            >ON</label
-                                        ></ng-template
-                                    >
-                                    <ng-template #elseBlock
-                                        ><label class="mb-0"
-                                            >OFF</label
-                                        ></ng-template
-                                    >
-                                </button>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-            <div
-                class="row"
                 *ngFor="let sidebarElement of sidebarElements">
                 <div
                     class="col-11 element-div ml-3 mt-3 text-nowrap"
